# server/generated/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "E-Commerce API Documentation",
    "version": "1.0.0",
    "description": "API Documentation for the E-Commerce backend (TypeScript) featuring authentication, user, product, cart, and order management."
  },
  "servers": [
    {
      "url": "http://localhost:5000",
      "description": "Development Server"
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Enter your JWT token in the format: Bearer <token>"
      }
    },
    "schemas": {
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "64f9b2c7e1234567890abcde"
          },
          "name": {
            "type": "string",
            "example": "John Doe"
          },
          "email": {
            "type": "string",
            "example": "john@example.com"
          },
          "role": {
            "type": "string",
            "enum": [
              "user",
              "admin"
            ],
            "example": "user"
          }
        },
        "required": [
          "id",
          "name",
          "email",
          "role"
        ]
      },
      "Product": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string",
            "example": "64f9b2c7e1234567890abcdf"
          },
          "name": {
            "type": "string",
            "example": "Mechanical Keyboard"
          },
          "description": {
            "type": "string",
            "example": "RGB mechanical keyboard"
          },
          "price": {
            "type": "number",
            "example": 99.99
          },
          "category": {
            "type": "string",
            "example": "Electronics"
          },
          "stock": {
            "type": "number",
            "example": 50
          },
          "imageUrl": {
            "type": "string",
            "example": "http://example.com/keyboard.jpg"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "_id",
          "name",
          "price",
          "category",
          "stock"
        ]
      },
      "CartItem": {
        "type": "object",
        "properties": {
          "product": {
            "$ref": "#/components/schemas/Product"
          },
          "quantity": {
            "type": "number",
            "minimum": 1,
            "example": 2
          }
        },
        "required": [
          "product",
          "quantity"
        ]
      },
      "Cart": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string",
            "example": "64f9b2c7e1234567890abcda"
          },
          "user": {
            "type": "string",
            "example": "64f9b2c7e1234567890abcde"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CartItem"
            }
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "_id",
          "user",
          "items"
        ]
      },
      "OrderItem": {
        "type": "object",
        "properties": {
          "product": {
            "$ref": "#/components/schemas/Product"
          },
          "quantity": {
            "type": "number",
            "minimum": 1,
            "example": 2
          },
          "price": {
            "type": "number",
            "example": 99.99
          }
        },
        "required": [
          "product",
          "quantity",
          "price"
        ]
      },
      "Order": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string",
            "example": "64f9b2c7e1234567890abcdb"
          },
          "user": {
            "type": "string",
            "example": "64f9b2c7e1234567890abcde"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OrderItem"
            }
          },
          "totalAmount": {
            "type": "number",
            "example": 199.98
          },
          "status": {
            "type": "string",
            "enum": [
              "pending",
              "processing",
              "shipped",
              "delivered",
              "cancelled"
            ],
            "example": "pending"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "_id",
          "user",
          "items",
          "totalAmount",
          "status"
        ]
      },
      "RegisterDTO": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "John Doe"
          },
          "email": {
            "type": "string",
            "format": "email",
            "example": "john@example.com"
          },
          "password": {
            "type": "string",
            "format": "password",
            "example": "password123"
          }
        },
        "required": [
          "name",
          "email",
          "password"
        ]
      },
      "LoginDTO": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "example": "john@example.com"
          },
          "password": {
            "type": "string",
            "format": "password",
            "example": "password123"
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "RefreshTokenDTO": {
        "type": "object",
        "properties": {
          "refreshToken": {
            "type": "string",
            "example": "eyJhbGciOi..."
          }
        },
        "required": [
          "refreshToken"
        ]
      },
      "AddCartItemDTO": {
        "type": "object",
        "properties": {
          "productId": {
            "type": "string",
            "example": "64f9b2c7e1234567890abcdf"
          },
          "quantity": {
            "type": "number",
            "minimum": 1,
            "example": 2
          }
        },
        "required": [
          "productId",
          "quantity"
        ]
      },
      "CreateProductDTO": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Mechanical Keyboard"
          },
          "description": {
            "type": "string",
            "example": "RGB mechanical keyboard"
          },
          "price": {
            "type": "number",
            "example": 99.99
          },
          "category": {
            "type": "string",
            "example": "Electronics"
          },
          "stock": {
            "type": "number",
            "example": 50
          },
          "imageUrl": {
            "type": "string",
            "example": "http://example.com/keyboard.jpg"
          }
        },
        "required": [
          "name",
          "price",
          "category"
        ]
      },
      "UpdateProductDTO": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Mechanical Keyboard"
          },
          "description": {
            "type": "string",
            "example": "RGB mechanical keyboard"
          },
          "price": {
            "type": "number",
            "example": 99.99
          },
          "category": {
            "type": "string",
            "example": "Electronics"
          },
          "stock": {
            "type": "number",
            "example": 50
          },
          "imageUrl": {
            "type": "string",
            "example": "http://example.com/keyboard.jpg"
          }
        }
      },
      "AuthResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "statusCode": {
            "type": "number",
            "example": 200
          },
          "message": {
            "type": "string",
            "example": "Login successful"
          },
          "data": {
            "type": "object",
            "properties": {
              "user": {
                "$ref": "#/components/schemas/User"
              },
              "accessToken": {
                "type": "string"
              },
              "refreshToken": {
                "type": "string"
              }
            },
            "required": [
              "user",
              "accessToken",
              "refreshToken"
            ]
          }
        },
        "required": [
          "success",
          "statusCode",
          "message",
          "data"
        ]
      },
      "RefreshResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "statusCode": {
            "type": "number",
            "example": 200
          },
          "message": {
            "type": "string",
            "example": "Access token refreshed successfully"
          },
          "data": {
            "type": "object",
            "properties": {
              "accessToken": {
                "type": "string"
              },
              "refreshToken": {
                "type": "string"
              }
            },
            "required": [
              "accessToken",
              "refreshToken"
            ]
          }
        },
        "required": [
          "success",
          "statusCode",
          "message",
          "data"
        ]
      },
      "ProductResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "statusCode": {
            "type": "number",
            "example": 200
          },
          "message": {
            "type": "string",
            "example": "Product fetched successfully"
          },
          "data": {
            "$ref": "#/components/schemas/Product"
          }
        },
        "required": [
          "success",
          "statusCode",
          "message",
          "data"
        ]
      },
      "ProductListResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "statusCode": {
            "type": "number",
            "example": 200
          },
          "message": {
            "type": "string",
            "example": "Products fetched successfully"
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Product"
            }
          }
        },
        "required": [
          "success",
          "statusCode",
          "message",
          "data"
        ]
      },
      "CartResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "statusCode": {
            "type": "number",
            "example": 200
          },
          "message": {
            "type": "string",
            "example": "Cart fetched successfully"
          },
          "data": {
            "$ref": "#/components/schemas/Cart"
          }
        },
        "required": [
          "success",
          "statusCode",
          "message",
          "data"
        ]
      },
      "OrderResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "statusCode": {
            "type": "number",
            "example": 201
          },
          "message": {
            "type": "string",
            "example": "Order placed successfully"
          },
          "data": {
            "$ref": "#/components/schemas/Order"
          }
        },
        "required": [
          "success",
          "statusCode",
          "message",
          "data"
        ]
      },
      "OrderListResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "statusCode": {
            "type": "number",
            "example": 200
          },
          "message": {
            "type": "string",
            "example": "Order history fetched successfully"
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Order"
            }
          }
        },
        "required": [
          "success",
          "statusCode",
          "message",
          "data"
        ]
      },
      "SuccessMessageResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "statusCode": {
            "type": "number",
            "example": 200
          },
          "message": {
            "type": "string",
            "example": "Operation successful"
          },
          "data": {
            "type": "object",
            "nullable": true,
            "default": null
          }
        },
        "required": [
          "success",
          "statusCode",
          "message",
          "data"
        ]
      },
      "DeleteProductResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "statusCode": {
            "type": "number",
            "example": 200
          },
          "message": {
            "type": "string",
            "example": "Product deleted successfully"
          },
          "data": {
            "type": "object",
            "properties": {
              "id": {
                "type": "string",
                "example": "64f9b2c7e1234567890abcdf"
              }
            },
            "required": [
              "id"
            ]
          }
        },
        "required": [
          "success",
          "statusCode",
          "message",
          "data"
        ]
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": false
          },
          "statusCode": {
            "type": "number",
            "example": 400
          },
          "message": {
            "type": "string",
            "example": "Error message details"
          }
        },
        "required": [
          "success",
          "statusCode",
          "message"
        ]
      }
    }
  },
  "paths": {
    "/api/users/register": {
      "post": {
        "tags": [
          "Users"
        ],
        "summary": "Register a new user",
        "description": "Register a new user with name, email, and password. Returns user info, accessToken, and refreshToken.",
        "operationId": "registerUser",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterDTO"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User registered successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request or user already exists",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/users/login": {
      "post": {
        "tags": [
          "Users"
        ],
        "summary": "User login",
        "description": "Authenticate user with email and password. Returns user info, accessToken, and refreshToken.",
        "operationId": "loginUser",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login successful",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid credentials",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/users/refresh-token": {
      "post": {
        "tags": [
          "Users"
        ],
        "summary": "Refresh access token",
        "description": "Get a new access token and refresh token pair using a valid refresh token.",
        "operationId": "refreshToken",
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshTokenDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Access token refreshed successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RefreshResponse"
                }
              }
            }
          },
          "401": {
            "description": "Invalid or expired refresh token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/users/logout": {
      "post": {
        "tags": [
          "Users"
        ],
        "summary": "Log out user",
        "description": "Revokes the refresh token and clears auth cookies.",
        "operationId": "logoutUser",
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshTokenDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Logged out successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessMessageResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/products": {
      "get": {
        "tags": [
          "Products"
        ],
        "summary": "Get all products",
        "description": "Retrieve a list of products, optionally filtered by category or search term.",
        "operationId": "getAllProducts",
        "parameters": [
          {
            "in": "query",
            "name": "category",
            "schema": {
              "type": "string"
            },
            "description": "Filter products by category name"
          },
          {
            "in": "query",
            "name": "search",
            "schema": {
              "type": "string"
            },
            "description": "Search query for product name or description"
          }
        ],
        "responses": {
          "200": {
            "description": "Products fetched successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductListResponse"
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "tags": [
          "Products"
        ],
        "summary": "Create a new product (Admin Only)",
        "description": "Adds a new product to the catalog.",
        "operationId": "createProduct",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProductDTO"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Product created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductResponse"
                }
              }
            }
          },
          "401": {
            "description": "Not authorized"
          },
          "403": {
            "description": "Forbidden (Admin only)"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
    "/api/products/{id}": {
      "get": {
        "tags": [
          "Products"
        ],
        "summary": "Get product by ID",
        "description": "Retrieves a single product by its MongoDB ObjectID.",
        "operationId": "getProductById",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "MongoDB identifier of the product"
          }
        ],
        "responses": {
          "200": {
            "description": "Product fetched successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductResponse"
                }
              }
            }
          },
          "404": {
            "description": "Product not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Server error"
          }
        }
      },
      "put": {
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "tags": [
          "Products"
        ],
        "summary": "Update product by ID (Admin Only)",
        "description": "Updates specified fields for an existing product.",
        "operationId": "updateProduct",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "MongoDB identifier of the product"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProductDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Product updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductResponse"
                }
              }
            }
          },
          "401": {
            "description": "Not authorized"
          },
          "403": {
            "description": "Forbidden (Admin only)"
          },
          "404": {
            "description": "Product not found"
          },
          "500": {
            "description": "Server error"
          }
        }
      },
      "delete": {
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "tags": [
          "Products"
        ],
        "summary": "Delete product by ID (Admin Only)",
        "description": "Removes a product from the database.",
        "operationId": "deleteProduct",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "MongoDB identifier of the product"
          }
        ],
        "responses": {
          "200": {
            "description": "Product deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteProductResponse"
                }
              }
            }
          },
          "401": {
            "description": "Not authorized"
          },
          "403": {
            "description": "Forbidden (Admin only)"
          },
          "404": {
            "description": "Product not found"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
    "/api/orders": {
      "get": {
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "tags": [
          "Orders"
        ],
        "summary": "Retrieve user's order history",
        "description": "Gets all orders placed by the authenticated user.",
        "operationId": "getOrderHistory",
        "responses": {
          "200": {
            "description": "Order history fetched successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderListResponse"
                }
              }
            }
          },
          "401": {
            "description": "Not authorized"
          },
          "500": {
            "description": "Server error"
          }
        }
      },
      "post": {
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "tags": [
          "Orders"
        ],
        "summary": "Place an order",
        "description": "Creates a new order using the current items in the user's cart, and clears the cart.",
        "operationId": "placeOrder",
        "responses": {
          "201": {
            "description": "Order placed successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request (e.g. cart is empty)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Not authorized"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
    "/api/cart": {
      "get": {
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "tags": [
          "Cart"
        ],
        "summary": "Retrieve user's cart",
        "description": "Gets the currently authenticated user's shopping cart.",
        "operationId": "getCart",
        "responses": {
          "200": {
            "description": "Cart fetched successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CartResponse"
                }
              }
            }
          },
          "401": {
            "description": "Not authorized"
          },
          "500": {
            "description": "Server error"
          }
        }
      },
      "post": {
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "tags": [
          "Cart"
        ],
        "summary": "Add item to cart",
        "description": "Adds a product with a specified quantity to the authenticated user's cart.",
        "operationId": "addItemToCart",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddCartItemDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Item added to cart successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CartResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request (e.g. invalid product ID or insufficient stock)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Not authorized"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    }
  },
  "tags": []
}